// lib/types/Profile.ts
import Departement from './depart'

type Profile = {
  firstName: string // Prénom
  lastName: string // Nom
  club: string // Club
  department: Departement // Département
  region: string // Région
  licenseNumber: string // Numéro de licence
  athletePage: string // Page athlète
}

export default Profile